+++ firmware/src/config/default/peripheral/tmr/plib_tmr5.h
@@ -77,11 +77,11 @@ void TMR5_Start(void);
 
 void TMR5_Stop(void);
 
-void TMR5_PeriodSet(uint16_t period);
+void TMR5_PeriodSet(uint32_t period);
 
-uint16_t TMR5_PeriodGet(void);
+uint32_t TMR5_PeriodGet(void);
 
-uint16_t TMR5_CounterGet(void);
+uint32_t TMR5_CounterGet(void);
 
 uint32_t TMR5_FrequencyGet(void);
