+++ firmware/src/config/default/peripheral/tmr/plib_tmr6.h
@@ -77,11 +77,11 @@ void TMR6_Start(void);
 
 void TMR6_Stop(void);
 
-void TMR6_PeriodSet(uint16_t period);
+void TMR6_PeriodSet(uint32_t period);
 
-uint16_t TMR6_PeriodGet(void);
+uint32_t TMR6_PeriodGet(void);
 
-uint16_t TMR6_CounterGet(void);
+uint32_t TMR6_CounterGet(void);
 
 uint32_t TMR6_FrequencyGet(void);
